Update of the file

+++ American-Pie.cs
@@ -6,10 +6,10 @@ namespace FractionSums
     {
         static void Main()
         {
-            long firstNominator = long.Parse(Console.ReadLine());
-            long firstDenominator = long.Parse(Console.ReadLine());
+            int firstNominator = long.Parse(Console.ReadLine());
+            int firstDenominator = long.Parse(Console.ReadLine());
 
-            long secondNominator = long.Parse(Console.ReadLine());
+            int secondNominator = long.Parse(Console.ReadLine());
             long secondDenominator = long.Parse(Console.ReadLine());
 
             long resultNominator = firstNominator * secondDenominator + secondNominator * firstDenominator;
